requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1
jsonschema>=4.23.0
tqdm>=4.66.4
streamlit>=1.36.0
pymupdf>=1.24.4
pillow>=10.3.0
pandas>=2.2.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
tiktoken>=0.5.2
numpy>=1.24.0
click>=8.1.7
rank-bm25>=0.2.2
nltk>=3.8

